Disable late acknowledgement for celery tasks
By default, messages are acknowledged right before being executed,
so they will not be redelivered if the worker fails.

With late acknowledgement, messages are not acknowledged until after
execution is completed, which causes the task to be retried if the
worker fails and can result in an infinite task loop if the task
itself is causing the worker failure.